+++ packages/nestjs-auth-github/src/auth-github.types.ts
@@ -1,0 +1,6 @@
+import { AuthGithubProfileInterface } from './interfaces/auth-github-profile.interface';
+import { AuthGithubCredentialsInterface } from './interfaces/auth-github-credentials.interface';
+
+export type MapProfile = (
+  profile: AuthGithubProfileInterface,
+) => AuthGithubCredentialsInterface;
