+++ packages/nestjs-auth-github/src/exceptions/auth-github-missing-email.exception.ts
@@ -4,7 +4,9 @@ export class AuthGithubMissingEmailException extends RuntimeException {
   constructor(
     message = 'GitHub did not return an email address for the user.',
   ) {
-    super(message);
+    super({
+      safeMessage: message,
+    });
     this.errorCode = 'AUTH_GITHUB_MISSING_PROFILE_EMAIL_ERROR';
   }
 }
