+++ packages/nestjs-auth-github/src/exceptions/auth-github-missing-id.exception.ts
@@ -2,7 +2,9 @@ import { RuntimeException } from '@concepta/nestjs-exception';
 
 export class AuthGithubMissingIdException extends RuntimeException {
   constructor(message = 'GitHub did not return an id for the user.') {
-    super(message);
+    super({
+      safeMessage: message,
+    });
     this.errorCode = 'AUTH_GITHUB_MISSING_PROFILE_ID_ERROR';
   }
 }
