+++ packages/nestjs-auth-github/src/index.ts
@@ -1,7 +1,7 @@
 export * from './auth-github.module';
 export * from './auth-github.controller';
 export * from './dto/auth-github-login.dto';
-export * from './interfaces/auth-github-map-profile.type';
+export * from './auth-github.types';
 
 export {
   AuthGithubGuard,
