+++ packages/nestjs-auth-github/src/interfaces/auth-github-credentials.interface.ts
@@ -1,11 +1,8 @@
 import {
+  ReferenceEmailInterface,
   ReferenceIdInterface,
-  ReferenceUsernameInterface,
 } from '@concepta/ts-core';
 
-/**
- * Credentials Interface
- */
 export interface AuthGithubCredentialsInterface
   extends ReferenceIdInterface,
-    ReferenceUsernameInterface {}
+    ReferenceEmailInterface {}
