--- packages/nestjs-auth-github/src/interfaces/auth-github-map-profile.type.ts
@@ -1,6 +1,0 @@
-import { AuthGithubProfileInterface } from './auth-github-profile.interface';
-import { AuthGithubSignInterface } from './auth-github-sign.interface';
-
-export type MapProfile = (
-  profile: AuthGithubProfileInterface,
-) => AuthGithubSignInterface;
