--- packages/nestjs-auth-github/src/interfaces/auth-github-sign.interface.ts
@@ -1,8 +1,0 @@
-import {
-  ReferenceEmailInterface,
-  ReferenceIdInterface,
-} from '@concepta/ts-core';
-
-export interface AuthGithubSignInterface
-  extends ReferenceIdInterface,
-    ReferenceEmailInterface {}
