+++ packages/nestjs-auth-google/src/auth-google.types.ts
@@ -1,0 +1,6 @@
+import { AuthGoogleProfileInterface } from './interfaces/auth-google-profile.interface';
+import { AuthGoogleSignInterface } from './interfaces/auth-google-credentials.interface';
+
+export type MapProfile = (
+  profile: AuthGoogleProfileInterface,
+) => AuthGoogleSignInterface;
