+++ packages/nestjs-auth-google/src/exceptions/auth-google-missing-email.exception.ts
@@ -4,7 +4,9 @@ export class AuthGoogleMissingEmailException extends RuntimeException {
   constructor(
     message = 'Google did not return an email address for the user.',
   ) {
-    super(message);
+    super({
+      safeMessage: message,
+    });
     this.errorCode = 'AUTH_GOOGLE_MISSING_PROFILE_EMAIL_ERROR';
   }
 }
