+++ packages/nestjs-auth-google/src/exceptions/auth-google-missing-id.exception.ts
@@ -2,7 +2,9 @@ import { RuntimeException } from '@concepta/nestjs-exception';
 
 export class AuthGoogleMissingIdException extends RuntimeException {
   constructor(message = 'Google did not return an id for the user.') {
-    super(message);
+    super({
+      safeMessage: message,
+    });
     this.errorCode = 'AUTH_GOOGLE_MISSING_PROFILE_ID_ERROR';
   }
 }
