+++ packages/nestjs-auth-google/src/interfaces/auth-google-credentials.interface.ts
@@ -1,11 +1,8 @@
 import {
+  ReferenceEmailInterface,
   ReferenceIdInterface,
-  ReferenceUsernameInterface,
 } from '@concepta/ts-core';
 
-/**
- * Credentials Interface
- */
-export interface AuthGoogleCredentialsInterface
+export interface AuthGoogleSignInterface
   extends ReferenceIdInterface,
-    ReferenceUsernameInterface {}
+    ReferenceEmailInterface {}
