--- packages/nestjs-auth-google/src/interfaces/auth-google-map-profile.type.ts
@@ -1,6 +1,0 @@
-import { AuthGoogleProfileInterface } from './auth-google-profile.interface';
-import { AuthGoogleSignInterface } from './auth-google-sign.interface';
-
-export type MapProfile = (
-  profile: AuthGoogleProfileInterface,
-) => AuthGoogleSignInterface;
