+++ packages/nestjs-auth-google/src/interfaces/auth-google-settings.interface.ts
@@ -1,6 +1,6 @@
 import { AuthenticationCodeInterface } from '@concepta/ts-common';
 import { Type } from '@nestjs/common';
-import { MapProfile } from './auth-google-map-profile.type';
+import { MapProfile } from '../auth-google.types';
 import { StrategyOptions } from 'passport-google-oauth20';
 
 export interface AuthGoogleSettingsInterface extends StrategyOptions {
