--- packages/nestjs-auth-google/src/interfaces/auth-google-sign.interface.ts
@@ -1,8 +1,0 @@
-import {
-  ReferenceEmailInterface,
-  ReferenceIdInterface,
-} from '@concepta/ts-core';
-
-export interface AuthGoogleSignInterface
-  extends ReferenceIdInterface,
-    ReferenceEmailInterface {}
