+++ packages/nestjs-auth-google/src/utils/auth-google-scope-parser.util.ts
@@ -1,12 +1,18 @@
+
 /**
- * Helper to split log level string and assign to correct log level type.
+ * Parses a comma-separated string of Google OAuth scopes into an array of individual scope strings.
  *
- * @internal
- * @param scopes - Log levels to split
+ * @param scopes - A comma-separated string of Google OAuth scopes.
+ * @returns An array of trimmed individual scope strings.
+ *
+ * @example
+ * const scopesString = "profile,email, openid";
+ * const parsedScopes = authGoogleScopeParse(scopesString);
+ * // Result: ["profile", "email", "openid"]
  */
-export function authGoogleScopeParse(scopes: string): string[] {
+export function authGoogleParseScope(scopes: string): string[] {
   // trim the string
-  const levelsTrimmed = scopes.trim();
+  const scopesTrimmed = scopes.trim();
 
   // is there any length?
   if (!scopes.length) {
@@ -15,11 +21,10 @@ export function authGoogleScopeParse(scopes: string): string[] {
   }
 
   // get raw strings
-  const levelTypes: string[] = levelsTrimmed.split(',');
+  const scopeTypes: string[] = scopesTrimmed.split(',');
 
-  // map all to log level enum
-  return levelTypes.map((levelType) => {
-    // trim the log level
-    return levelType.trim();
+  // trim all scopes
+  return scopeTypes.map((scope) => {
+    return scope.trim();
   });
 }
